\name{consrq-package}
\alias{consrq-package}
\docType{package}
\title{
Constrained Quantile Regression
}

\description{
Constrained quantile regression is performed. One constraint is that all beta coefficients
(including the constant) cannot be negative. They can be either 0 or strictly positive.
Another constraint is that the beta coefficients lie within an interval.
}

\details{
\tabular{ll}{
Package: \tab consrq\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2024-11-20 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Koenker R. (2005) Quantile Regression, Cambridge University Press.
}





